mediatek: apsoc: move `plat_is_my_cpu_primary` into assembly helpers
This function is called before the C stack is set up, which resulted
in hangs when TF‐A is compiled with `-fno-omit-frame-pointer`,
as is now the default on several major Linux distributions.

As a slightly broken non‐stub implementation was already available
for the equivalent obsolete `platform_is_primary_cpu` API, fix and
rename it and drop the C functions.

If this logic is truly unnecessary, then `COLD_BOOT_SINGLE_CPU`
could be set instead and these functions removed.

Fixes: 6ea8b646e063e6f848deb4ef30814e41ecfbb55c ("mediatek: mt7622: add initial BL2/BL31 support")
Fixes: 36bebd65bf9f2af3504afbc6f8cd2c2e650be66b ("mediatek: mt7986: add initial BL2/BL31 support")
Fixes: fd8aecae81c0a34200bdaf3e75d7bdc3cf39070a ("mediatek: mt7981: add initial BL2/BL31 support")
Fixes: ced5322fb549e978057b09ef0ef529a893df74bc ("mediatek: mt7988: add initial BL2/BL31 support")
Fixes: e838b234347fa1149913c5cacf04ba8ca15cc2fa ("mediatek: mt7987: add initial BL2/BL31 support")
Signed-off-by: Emily <hello@emily.moe>